Fix bug: colored headers with noWhiteSpace incorrectly spaced
Prior to this change, we'd incorrectly space out headers

Date                   Name               Email
2024-09-19T22:16:24Z  pigwantacat       refactor:refactor RandomString

can see that the header has an extra space with each column.

The issue is that on line 635 in the old code (637 new), we were putting
a space between the padded header and the column padding. Removing that
space seems to fix it, and doesn't break any unit testing (which is a
sign that this wasn't being tested), so probably an oversight in the
original implementation.

I think this bug was introduced in commit 50793076 .

After this change, the above example correctly formats:

Date                  Name              Email
2024-09-19T22:16:24Z  pigwantacat       refactor:refactor RandomString