Replace custom color handling with fatih/color
The custom color handling in tablewriter is fairly basic and mostly does
the job, but for Rad, we need to be able to redirect output to e.g. a
file. However, with this implementation, the redirected output includes
the ANSI color codes, which is undesirable.

Let's replace the coloring implementation with fatih/color which has a
nice API and also handles turning off coloring when needed e.g. for
redirection.