Support multiple coloring column mods
The previous commit allowed only one coloring mod in a column. We want
to support configuring several.

The handling is quite complex, so it's heavily commented. The complexity
arises from correctly handling overlapping regexes, and which to prio,
particularly because earlier passes, in implemented naively, will
prevent later regexes from matching due to the insertion of ANSI escape
codes.